Pull EntityTile body text into own component
