+++ context/app/static/js/components/entity-tile/EntityTile/EntityTile.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
-import Typography from '@material-ui/core/Typography';
 import format from 'date-fns/format';
+
+import EntityTileBodyText from '../EntityTileBodyText';
 
 import {
   StyledDatasetIcon,
@@ -9,10 +10,6 @@ import {
   StyledPaper,
   HoverOverlay,
   FixedWidthFlex,
-  Flex,
-  TruncatedTypography,
-  StyledDivider,
-  TextSection,
   BottomSection,
   BottomSectionDivider,
   BottomSectionText,
@@ -31,30 +28,12 @@ function EntityTile(props) {
             {entity_type === 'Dataset' && <StyledDatasetIcon $isCurrentEntity={isCurrentEntity} />}
             {entity_type === 'Donor' && <StyledDonorIcon $isCurrentEntity={isCurrentEntity} />}
             {entity_type === 'Sample' && <StyledSampleIcon $isCurrentEntity={isCurrentEntity} />}
-            <TextSection>
-              <Typography component="h4" variant="h6">
-                {id}
-              </Typography>
-              {'origin_sample' in entityData && (
-                <TruncatedTypography variant="body2">{entityData.origin_sample.mapped_organ}</TruncatedTypography>
-              )}
-              {'mapped_specimen_type' in entityData && (
-                <TruncatedTypography variant="body2">{entityData.mapped_specimen_type}</TruncatedTypography>
-              )}
-              {'mapped_data_types' in entityData && (
-                <TruncatedTypography variant="body2">{entityData.mapped_data_types.join(', ')}</TruncatedTypography>
-              )}
-              {entity_type === 'Donor' && 'mapped_metadata' in entityData && (
-                <>
-                  <Flex>
-                    <Typography variant="body2">{entityData.mapped_metadata.gender}</Typography>
-                    <StyledDivider flexItem orientation="vertical" $isCurrentEntity={isCurrentEntity} />
-                    <Typography variant="body2">{entityData.mapped_metadata.age} years</Typography>
-                  </Flex>
-                  <TruncatedTypography variant="body2">{entityData.mapped_metadata.race}</TruncatedTypography>
-                </>
-              )}
-            </TextSection>
+            <EntityTileBodyText
+              entity_type={entity_type}
+              id={id}
+              isCurrentEntity={isCurrentEntity}
+              entityData={entityData}
+            />
           </FixedWidthFlex>
           <BottomSection $isCurrentEntity={isCurrentEntity}>
             {Object.entries(descendantCounts).map(([k, v]) => (
