+++ context/app/static/js/components/entity-tile/EntityTileBodyText/index.js
@@ -1,0 +1,3 @@
+import EntityTileBodyText from './EntityTileBodyText';
+
+export default EntityTileBodyText;
